Create cancer forecasting
The inside of a tumour can be visualized on a computer by using mathematical models. Image: ALEXANDER R. A. ANDERSON, WWW.MOFFITT.ORG/IMO
Mathematicians are making inroads in oncology with approaches that used to be the realm of biologists: building intricate models of cancer ecology and evolution. It's a logical transition. The forces of environmental selection and adaptation that act on organisms also govern the way cancer cells develop, and a better understanding of how the disease evolves and adapts to environmental changes can help researchers find a way to stop it. “Evolutionary theory is the theory for cancer,” says Carlo Maley, a computational and evolutionary biologist at the University of California, San Francisco. “If we can engage with this evolutionary process, we can slow it down or shunt it in a direction that's more manageable.”