Update README.md
This detailed ability to predict invasiveness is particularly important in diseases such as prostate cancer, in which some tumours are removed unnecessarily and treated aggressively. To assess the model's predictive accuracy, Anderson's team, working with a biologist, is testing it in mice. If the model works, it would be the first to predict a tumour's aggression before the start of treatment, providing oncologists with a guide for treatment options.
Beyond evolutionary dynamics, mathematical models are helping to reshape the way some researchers approach cancer treatments. The improved understanding of cancer evolution provided by the models provide has led some researchers to a counter-intuitive conclusion: for most cancers, targeted therapy just doesn't work. When you use drugs that target specific mutations, Anderson says, “our models tell us that you're destined to see drug resistance because there's another route to the same phenotype”. Cells that survive the first wave of therapy can evolve to have the same malignant characteristics but without the same vulnerability to the drug. What's more, the evolutionary dynamics of tumours suggest that trying to eliminate cancers speeds up the development of drug resistance, reducing a patient's chances of survival. Models show that there might be a better way.
Changing the game
Some researchers are looking for ways to hack cancer's programming and turn it against itself. Robert Gatenby, an oncologist at the Moffitt Cancer Center, has been seeking fundamental laws he can exploit. One way to attack tumours is to make them similar to each other; the more homogenous they are, the lower the odds that they will become resistant to treatment, and the greater the chances that therapy will be effective. In essence, he is looking for tricks to prevent the evolution of drug resistance.